Clean up Water, Not perfect but better
The code half works, in that it generates a somewhat desirable terrain but there are inconsistencies in the output. Sometimes random cliff tiles are created in the center of a body of water. The bigger issue is that is only places cliffs or edges due to the raw noise value which is equal in areas. So some form of intervention is needed to have cliffs and ramps mixed for instances of elevation layers. The rotation of the tiles is also incorrect and needs work to rotate correctly, which is assumed to be an easy fix to do later.