Block out new Version
The original version of the code may have produced a terrain, but a terrain is more than just a piece of land. It needs walkable paths/roads, areas for plant life like trees, open areas for large structures like buildings or cities. The new version aims to meet the needs of a terrain by providing the ability to procedurally generate natural elements at the core.