Draft a new design to better approach the problem
Previously the code manipulated cells and changed cells that were changed by other functions. This led to creation of water systems and large open area's encroaching on each other's land. The smooth operations were expanding and affecting cells that shouldn't be changed. The new code is setup to be able to work towards a more flow system of determining when a portion of the grid is deemed final.